Disconnected the passwords for mpd and web from mysql.
The password for mpd an web were set to the password of mysql.
Since I wouldn't like the passwords to be the same, I chose to
make it more obvious that they could be different. Of course must
the mpd password be the same as the one set in the mpd.conf